fixed "Unexpected value type of 'memory'."
I was trying out your tutorial and I got "The request content was malformed. Unexpected value type of 'memory'" trying to create the first blueprint. Digging around in the yaml here it seems like the "64MB" on the docs doesn't match the "256" shown in the latest yaml. So here is the fix.